help typecheck
