+++ packages/core/src/generation.ts
@@ -1,7 +1,6 @@
 // Import necessary modules and interfaces
-import { CancellationToken } from "./cancellation"
-import { LanguageModel } from "./chat"
-import {
+import type { CancellationToken } from "./cancellation"
+import type {
     ChatCompletionMessageParam,
     ChatCompletionsOptions,
     ChatCompletionUsage,
@@ -13,84 +12,6 @@ import { GenerationStats } from "./usage"
 export interface Fragment {
     files: string[] // Array of file paths or names
 }
-
-// Interface for the result of a generation process
-export interface GenerationResult extends GenerationOutput {
-    /**
-     * The environment variables passed to the prompt
-     */
-    vars: Partial<ExpansionVariables>
-
-    /**
-     * Expanded prompt text composed of multiple messages
-     */
-    messages: ChatCompletionMessageParam[]
-
-    /**
-     * Edits to apply, if any
-     */
-    edits: Edits[]
-
-    /**
-     * Source annotations parsed as diagnostics
-     */
-    annotations: Diagnostic[]
-
-    /**
-     * Sections of the ChangeLog
-     */
-    changelogs: string[]
-
-    /**
-     * Error message or object, if any error occurred
-     */
-    error?: unknown
-
-    /**
-     * Status of the generation process (success, error, or cancelled)
-     */
-    status: GenerationStatus
-
-    /**
-     * Additional status information or message
-     */
-    statusText?: string
-
-    /**
-     * Completion status from the language model
-     */
-    finishReason?: string
-
-    /**
-     * Optional label for the run
-     */
-    label?: string
-
-    /**
-     * Version of the GenAIScript used
-     */
-    version: string
-
-    /**
-     * Logprobs if computed
-     */
-    logprobs?: Logprob[]
-
-    /**
-     * Statistics of the generation
-     */
-    perplexity?: number
-
-    /**
-     * Statistics of the generation
-     */
-    stats?: {
-        cost: number
-    } & ChatCompletionUsage
-}
-
-// Type representing possible statuses of generation
-export type GenerationStatus = "success" | "error" | "cancelled" | undefined
 
 // Options for configuring the generation process, extending multiple other options
 export interface GenerationOptions
