+++ packages/core/src/promptrunner.ts
@@ -1,15 +1,15 @@
 // Import necessary modules and functions for handling chat sessions, templates, file management, etc.
 import { executeChatSession, tracePromptResult } from "./chat"
-import { Project } from "./server/messages"
+import { GenerationStatus, Project } from "./server/messages"
 import { arrayify, assert, relativePath } from "./util"
 import { runtimeHost } from "./host"
 import { MarkdownTrace } from "./trace"
 import { CORE_VERSION } from "./version"
 import { expandFiles } from "./fs"
 import { CSVToMarkdown } from "./csv"
-import { Fragment, GenerationOptions, GenerationStatus } from "./generation"
+import { Fragment, GenerationOptions } from "./generation"
 import { traceCliArgs } from "./clihelp"
-import { GenerationResult } from "./generation"
+import { GenerationResult } from "./server/messages"
 import { resolveModelConnectionInfo } from "./models"
 import { RequestError, errorMessage } from "./error"
 import { renderFencedVariables } from "./fence"
