+++ packages/core/src/server/messages.ts
@@ -1,6 +1,5 @@
-import { ChatCompletionAssistantMessageParam } from "../chattypes"
-import { GenerationResult } from "../generation"
-import { LanguageModelConfiguration, ResponseStatus } from "../host"
+import type { ChatCompletionAssistantMessageParam, ChatCompletionMessageParam, ChatCompletionUsage } from "../chattypes"
+import type { LanguageModelConfiguration, ResponseStatus } from "../host"
 
 /**
  * Represents a project containing templates and diagnostics.
@@ -121,6 +120,84 @@ export interface PromptScriptStart extends RequestMessage {
 
 export interface PromptScriptStartResponse extends ResponseStatus {
     runId: string
+}
+
+// Type representing possible statuses of generation
+export type GenerationStatus = "success" | "error" | "cancelled" | undefined
+
+// Interface for the result of a generation process
+export interface GenerationResult extends GenerationOutput {
+    /**
+     * The environment variables passed to the prompt
+     */
+    vars: Partial<ExpansionVariables>
+
+    /**
+     * Expanded prompt text composed of multiple messages
+     */
+    messages: ChatCompletionMessageParam[]
+
+    /**
+     * Edits to apply, if any
+     */
+    edits: Edits[]
+
+    /**
+     * Source annotations parsed as diagnostics
+     */
+    annotations: Diagnostic[]
+
+    /**
+     * Sections of the ChangeLog
+     */
+    changelogs: string[]
+
+    /**
+     * Error message or object, if any error occurred
+     */
+    error?: unknown
+
+    /**
+     * Status of the generation process (success, error, or cancelled)
+     */
+    status: GenerationStatus
+
+    /**
+     * Additional status information or message
+     */
+    statusText?: string
+
+    /**
+     * Completion status from the language model
+     */
+    finishReason?: string
+
+    /**
+     * Optional label for the run
+     */
+    label?: string
+
+    /**
+     * Version of the GenAIScript used
+     */
+    version: string
+
+    /**
+     * Logprobs if computed
+     */
+    logprobs?: Logprob[]
+
+    /**
+     * Statistics of the generation
+     */
+    perplexity?: number
+
+    /**
+     * Statistics of the generation
+     */
+    stats?: {
+        cost: number
+    } & ChatCompletionUsage
 }
 
 export interface PromptScriptEndResponseEvent {
