+++ packages/vscode/src/state.ts
@@ -22,9 +22,10 @@ import { isCancelError } from "../../core/src/error"
 import { MarkdownTrace } from "../../core/src/trace"
 import { logInfo, groupBy } from "../../core/src/util"
 import { CORE_VERSION } from "../../core/src/version"
-import { Fragment, GenerationResult } from "../../core/src/generation"
+import { GenerationResult } from "../../core/src/server/messages"
 import { hash, randomHex } from "../../core/src/crypto"
 import { delay } from "es-toolkit"
+import { Fragment } from "../../core/src/generation"
 
 export const FRAGMENTS_CHANGE = "fragmentsChange"
 export const AI_REQUEST_CHANGE = "aiRequestChange"
